done refactoring #11
